+++ packages/components-react/tests/e2e/specs/examples.e2e.ts
@@ -1,24 +1,13 @@
-import type { Page } from 'puppeteer';
+import { test, expect } from '@playwright/test';
 import { goto, initConsoleObserver, getConsoleErrorsAmount, getConsoleWarningsAmount } from '../helpers';
 import * as fs from 'fs';
 import * as path from 'path';
-import { createRequire } from 'node:module';
 
-// @ts-ignore
-const nodeRequire = createRequire(import.meta.url);
-
-let page: Page;
-
-beforeEach(async () => {
-  page = await browser.newPage();
+test.beforeEach(async ({ page }) => {
   initConsoleObserver(page);
 });
-afterEach(async () => await page.close());
 
-const filePath = path.resolve(
-  nodeRequire.resolve('@porsche-design-system/components-react'),
-  '../../../../src/routes.tsx'
-);
+const filePath = path.resolve(require.resolve('@porsche-design-system/components-react'), '../../../../src/routes.tsx');
 const fileContent = fs.readFileSync(filePath, 'utf8');
 
 const [, rawRoutes] = /const routes.*(\[[\s\S]*\]);/.exec(fileContent) || [];
@@ -31,14 +20,16 @@ const routes: { name: string; path: string; element: string }[] = eval(
 const exampleRoutes = routes.filter((item) => item.element.startsWith('fromExamples.'));
 const exampleUrls = exampleRoutes.map((item) => item.path.slice(1));
 
-it.each(exampleUrls)('should work without error or warning for %s', async (exampleUrl) => {
-  await goto(page, exampleUrl);
-  expect(getConsoleErrorsAmount()).toBe(0);
-  expect(getConsoleWarningsAmount()).toBe(0);
+for (const exampleUrl of exampleUrls) {
+  test(`should work without error or warning for ${exampleUrl}`, async ({ page }) => {
+    await goto(page, exampleUrl);
+    expect(getConsoleErrorsAmount()).toBe(0);
+    expect(getConsoleWarningsAmount()).toBe(0);
 
-  await page.evaluate(() => console.error('test error'));
-  expect(getConsoleErrorsAmount()).toBe(1);
+    await page.evaluate(() => console.error('test error'));
+    expect(getConsoleErrorsAmount()).toBe(1);
 
-  await page.evaluate(() => console.warn('test warning'));
-  expect(getConsoleWarningsAmount()).toBe(1);
-});
+    await page.evaluate(() => console.warn('test warning'));
+    expect(getConsoleWarningsAmount()).toBe(1);
+  });
+}