+++ packages/components-react/tests/e2e/specs/testing.e2e.ts
@@ -1,6 +1,7 @@
+import { test, expect } from '@playwright/test';
 import * as jsTesting from '@porsche-design-system/components-js/testing';
 import * as reactTesting from '@porsche-design-system/components-react/testing';
 
-it('should reexport testing sub-package from components-js', () => {
+test('should reexport testing sub-package from components-js', () => {
   expect(reactTesting).toEqual(jsTesting);
 });