+++ packages/components/src/styles/flyout-dialog-reset-styles.ts
@@ -3,8 +3,8 @@ import { JssStyle } from 'jss';
 export const getFlyoutDialogResetJssStyle = (): JssStyle => {
   return {
     position: 'fixed',
-    height: '100vh',
-    maxHeight: '100vh',
+    height: '100dvh',
+    maxHeight: '100dvh',
     margin: '0',
     padding: '0',
     border: '0',