+++ packages/crawler/src/index.ts
@@ -1,5 +1,4 @@
-import * as puppeteer from 'puppeteer';
-import * as puppeteerConfig from '@porsche-design-system/shared/testing/jest-puppeteer.config';
+import { chromium } from 'playwright';
 import { crawlWebsites } from './crawl-websites';
 import { crawlerConfig } from '../constants';
 
@@ -13,15 +12,14 @@ const startBrowser = async (): Promise<void> => {
     'https://footer.pchomenav.aws.porsche.cloud/react/index.html?locale=de',
     'https://porsche.com/germany/aboutporsche/jobs',
     'https://porsche.com/germany/aboutporsche/porschemuseum',
-    'https://porsche.com/usa/models',
-    'https://porsche.com/usa/modelstart/all',
-    'https://porsche.com/usa/compare/?model=982120&modelToCompare=982320&modelGearType=480&modelToCompareGearType=480',
-    'https://porsche.com/usa/compare/?modelRange=718',
+    'https://porsche.com/germany/models',
+    'https://porsche.com/germany/modelstart/all',
+    'https://porsche.com/germany/compare/?model=982120&modelToCompare=982320&modelGearType=480&modelToCompareGearType=480',
+    'https://porsche.com/germany/compare/?modelRange=718',
     'https://configurator.porsche.com/de-DE/model/992110',
     'https://porsche.com/stories',
     'https://dealer.porsche.com/ch/zug/de-CH',
     'https://dealer.porsche.com/ch/zug/de-CH/neuwagen/modelle',
-    'https://login.porsche.com/login/de/de_DE',
     'https://signup.porsche.com/gb/en_GB/register',
     'https://finder.porsche.com/de/de-DE',
     'https://finder.porsche.com/de/de-DE/search',
@@ -33,27 +31,32 @@ const startBrowser = async (): Promise<void> => {
     'https://shop.porsche.com/de/de-DE/checkout?po-id=84460f8a-c9bf-434d-9dc5-398a55940995',
     'https://connect-store.porsche.com/de/de?model=911_2022',
     'https://connect-store.porsche.com/de/de/p/bundle_connect_v1?model=911_2022',
+    'https://connect-store.porsche.com/offer/de/de-DE/macan_2024/products',
+    'https://connect-store.porsche.com/offer/de/de-DE/macan_2024/products/bundle_connect_v4',
   ];
 
-  const browser = await puppeteer.launch({
-    ...puppeteerConfig,
-    ...{
+  const browser = await chromium.launch({
+    headless: true,
+    args: [
+      '--no-sandbox',
+      '--disable-setuid-sandbox',
+      '--disable-dev-shm-usage',
+      '--single-process',
+      '--disable-web-security',
       // since some websites have different components depending on the window size & viewport,
       // we have to set the window size and viewport here in addition to shared config
-      args: [`--window-size=${crawlerConfig.viewport.width},${crawlerConfig.viewport.height}`],
-      defaultViewport: {
-        width: crawlerConfig.viewport.width,
-        height: crawlerConfig.viewport.height,
-      },
-    },
+      `--window-size=${crawlerConfig.viewport.width},${crawlerConfig.viewport.height}`,
+    ],
   });
   console.log('Crawling websites..');
-  // crawling all websites
+
   await crawlWebsites(browser, customerWebsites);
+
   console.log('Success - please check reports');
+
   await browser.close();
 };
 
 startBrowser().catch((err) => {
-  throw new Error(`An error occurred during crawling: ${err}`);
+  throw err;
 });