+++ packages/components/CHANGELOG.md
@@ -14,6 +14,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### [Unreleased]
 
+### [3.13.1] - 2024-03-20
+
+#### Fixed
+
+- `Modal`: Unexpected scrolling behavior on iOS >= 17.4
+  ([#3128](https://github.com/porsche-design-system/porsche-design-system/pull/3128))
+- `Select`, `Multi-Select`: Cropping issues of select dropdown when used inside `Table` component
+  ([#3114](https://github.com/porsche-design-system/porsche-design-system/pull/3114))
+- `Flyout`, `Flyout Navigation`: iOS Safari URL bar overlaying
+  ([#3131](https://github.com/porsche-design-system/porsche-design-system/pull/3131))
+
 ### [3.13.0] - 2024-03-11
 
 ### [3.13.0-rc.2] - 2024-03-11